+++ src/app-ui/index.js
@@ -3,7 +3,7 @@ var schemas = require('ssb-msg-schemas')
 
 module.exports = function (phoenix) {
 
-  require('./invite-flow')(phoenix)
+  require('./modals')(phoenix)
 
   phoenix.ui.showUserId = function () { 
     swal('Here is your contact id', phoenix.user.id)
@@ -18,54 +18,6 @@ module.exports = function (phoenix) {
   phoenix.ui.enableScrolling = function () {
     document.querySelector('html').style.overflow = 'auto'
     window.scrollTo(0, oldScrollTop)
-  }
-
-  phoenix.ui.modal = function (el) {
-    // create a context so we can release this modal on close
-    var h2 = h.context()
-
-    // markup
-
-    var inner = h2('.modal-inner', el)
-    var modal = h2('.modal', { onclick: onmodalclick }, inner)
-    document.body.appendChild(modal)
-
-    modal.close = function () {
-      // check if there are any forms in progress
-      var els = Array.prototype.slice.call(inner.querySelectorAll('textarea'))
-      for (var i=0; i < els.length; i++) {
-        if (els[i].value) {
-          if (!confirm('Close modal and lose changes to your reply?'))
-            return
-          break
-        }
-      }
-
-      // remove
-      document.body.removeChild(modal)
-      window.removeEventListener('hashchange', modal.close)
-      window.removeEventListener('keyup', onkeyup)
-      h2.cleanup()
-      phoenix.ui.enableScrolling()
-      modal = null
-    }
-
-    // handlers
-
-    function onmodalclick (e) {
-      if (e.target == modal)
-        modal.close()
-    }
-    function onkeyup (e) {
-      // close on escape
-      if (e.which == 27)
-        modal.close()
-    }
-    window.addEventListener('hashchange', modal.close)
-    window.addEventListener('keyup', onkeyup)
-    phoenix.ui.disableScrolling()
-
-    return modal
   }
 
   phoenix.ui.setStatus = function (type, message) {
