+++ src/com/index.js
@@ -186,7 +186,7 @@ exports.sidehelp = function (app, opts) {
   return h('ul.list-unstyled.sidehelp',
     h('li', h('button.btn.btn-link', { onclick: app.ui.showUserId }, 'Get your id')),
     h('li', h('button.btn.btn-link', { onclick: app.ui.followPrompt }, 'Add a contact')),
-    h('li', h('button.btn.btn-link', { onclick: app.ui.inviteFlow }, 'Use an invite')),
+    h('li', h('button.btn.btn-link', { onclick: app.ui.inviteModal }, 'Use an invite')),
     (!opts || !opts.noMore) ? h('li', h('span', {style:'display: inline-block; padding: 6px 14px'}, a('#/help', 'More help'))) : ''
   )
 }
